动态路径计算更新

+++ src/main/java/cn/edu/sicau/pfdistribution/dao/mysqlsave/SelectLineid.java
@@ -1,13 +1,13 @@
 package cn.edu.sicau.pfdistribution.dao.mysqlsave;
 
 import lombok.Data;
-import java.math.BigDecimal;
+
 
 @Data
 public class SelectLineid {
-    public BigDecimal LINE_ID;
+    public int LINE_ID;
     public String LINE_NAME;
-    public BigDecimal CZ_SN;
-    public BigDecimal CZ_ID;
+    public int CZ_SN;
+    public int CZ_ID;
     public String CZ_NAME;
 }
